Cleanup Il2cpp stuff
