+++ Drova Modding API/Systems/Spawning/LazyActorSaveDataManaged.cs
@@ -1,0 +1,20 @@
+﻿
+public class LazyActorSaveDataManaged
+{
+    public string ActorName = "";
+    public string ActorReferenceString = "";
+    public string ActorEnitityInfoReferenceString = "";
+    public string ActorGuid = "";
+
+    public LazyActorSaveDataManaged()
+    {
+    }
+
+    public LazyActorSaveDataManaged(string actorName, string actorReferenceString, string actorEnitityInfoReferenceString, string actorGuid)
+    {
+        ActorName = actorName;
+        ActorReferenceString = actorReferenceString;
+        ActorEnitityInfoReferenceString = actorEnitityInfoReferenceString;
+        ActorGuid = actorGuid;
+    }
+}
